saturn fdt: Fix scripts according shellcheck.

+++ board/ntlab/saturn_fdt/flash_emmc_linux.sh
@@ -20,7 +20,7 @@ mkdir -p $dst_part
 #mkdir /tmp/src_image
 mount /dev/mmcblk1p1 $dst_part
 #mount /dev/mmcblk0p2 /tmp/src_image
-rm -rf $dst_part/*
+rm -rf "${dst_part:?}"/*
 
 #tar xf /tmp/src_image/rootfs.tar -C /tmp/dst_part
 tar xf rootfs.tar -C $dst_part
